Cleanup bunch of MSVC compiler warnings
<==== Mercurial data ====>
Branch: dev_3.0.0_gui

+++ client/libclient/src/core/resource/resource_directory_browser.cpp
@@ -269,7 +269,7 @@ QnLayoutResourcePtr QnResourceDirectoryBrowser::layoutFromFile(const QString& xf
             {
                 NX_ASSERT(motionIO->size() % sizeof(QnMetaDataV1Light) == 0);
                 QnMetaDataLightVector motionData;
-                int motionDataSize = motionIO->size() / sizeof(QnMetaDataV1Light);
+                size_t motionDataSize = motionIO->size() / sizeof(QnMetaDataV1Light);
                 if (motionDataSize > 0)
                 {
                     motionData.resize(motionDataSize);
